🧪 NWJCBW task: split config policy coverage
